added RecommendedWords.js component
